+++ poteto.go
@@ -1,17 +1,18 @@
 package poteto
 
 import (
-	"bytes"
 	"net/http"
-	"os"
 	"strings"
 	"sync"
 
 	"github.com/fatih/color"
 	"github.com/poteto0/poteto/constant"
+	"github.com/poteto0/poteto/utils"
 )
 
 type Poteto interface {
+	// If requested, call this
+	// you can make WithRequestIdOption false: you can faster request
 	ServeHTTP(w http.ResponseWriter, r *http.Request)
 	Run(addr string)
 	GET(path string, handler HandlerFunc) error
@@ -29,6 +30,7 @@ type poteto struct {
 	middlewareTree MiddlewareTree
 	logger         any
 	cache          sync.Pool
+	option         PotetoOption
 }
 
 func New() Poteto {
@@ -36,6 +38,16 @@ func New() Poteto {
 		router:         NewRouter([]string{"GET", "POST", "PUT", "DELETE"}),
 		errorHandler:   &httpErrorHandler{},
 		middlewareTree: NewMiddlewareTree(),
+		option:         DefaultPotetoOption,
+	}
+}
+
+func NewWithOption(option PotetoOption) Poteto {
+	return &poteto{
+		router:         NewRouter([]string{"GET", "POST", "PUT", "DELETE"}),
+		errorHandler:   &httpErrorHandler{},
+		middlewareTree: NewMiddlewareTree(),
+		option:         option,
 	}
 }
 
@@ -56,12 +68,13 @@ func (p *poteto) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// get from cache & reset context
 	ctx := p.initializeContext(w, r)
 
-	// get and SetRequestId
-	// slow -> should be middleware?
-	reqId := ctx.RequestId()
-	ctx.Set(constant.STORE_REQUEST_ID, reqId)
-	if id := ctx.GetRequest().Header.Get(constant.HEADER_X_REQUEST_ID); id == "" {
-		ctx.GetResponse().Header().Set(constant.HEADER_X_REQUEST_ID, reqId)
+	// default get & set RequestId
+	if p.option.WithRequestId {
+		reqId := ctx.RequestId()
+		ctx.Set(constant.STORE_REQUEST_ID, reqId)
+		if id := ctx.GetRequest().Header.Get(constant.HEADER_X_REQUEST_ID); id == "" {
+			ctx.GetResponse().Header().Set(constant.HEADER_X_REQUEST_ID, reqId)
+		}
 	}
 
 	routes := p.router.GetRoutesByMethod(r.Method)
@@ -98,18 +111,14 @@ func (p *poteto) applyMiddleware(middlewares []MiddlewareFunc, handler HandlerFu
 
 func (p *poteto) Run(addr string) {
 	// Print Banner
-	buf := &bytes.Buffer{}
 	coloredBanner := color.HiGreenString(Banner)
-	buf.WriteString(coloredBanner)
-	buf.WriteTo(os.Stdout)
+	utils.PotetoPrint(coloredBanner)
 
 	if !strings.Contains(addr, constant.PARAM_PREFIX) {
 		addr = constant.PARAM_PREFIX + addr
 	}
 
-	msg := "serve at http://localhost:" + addr + "\n"
-	buf.WriteString(msg)
-	buf.WriteTo(os.Stdout)
+	utils.PotetoPrint("server is available at http://localhost" + addr + "\n")
 
 	if err := http.ListenAndServe(addr, p); err != nil {
 		panic(err)