+++ poteto_option.go
@@ -1,0 +1,9 @@
+package poteto
+
+type PotetoOption struct {
+	WithRequestId bool `yaml:"with_request_id"`
+}
+
+var DefaultPotetoOption = PotetoOption{
+	WithRequestId: true,
+}