+++ utils/output.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"bytes"
+	"os"
+)
+
+func PotetoPrint(msg string) {
+	buf := bytes.NewBuffer([]byte(msg))
+	buf.WriteTo(os.Stdout)
+}